move Apache POI related tests to separate class (ad. #58)
